fix apk build with JDK 21 and above
Older versions of the Android Gradle Plugin (AGP) are apparently
incompatible with JDK 21 and above [1]. Thus, update AGP to 8.10.2 and
relevant plugins as well, as described in [2,3]. Furthermore, to avoid
the warning shown in [4], `sourceCompatibility`, `targetCompatibility`
and the kotlin `jvmTarget` need to be updated as well. According to [3],
JDK 17 currently works best with the latest Gradle and Android plugin
versions.

[1] https://github.com/flutter/flutter/issues/156304
[2] https://github.com/flutter/flutter/issues/156304#issuecomment-2397707812
[3] https://stackoverflow.com/a/79328913
[4] https://stackoverflow.com/q/79102777